linkedlist in C answar

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,6 +1,6 @@
 #include"lists.h"
-#include<stdio.h>
-#include<stdlib.h>
+
+
 /**
  * print_listint - func to print linked list
  * @h: pointer for header
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int count = 0;
+	size_t count = 0;
 	listint_t *temp;
 
 	for (; h != NULL; count++)
